Migrate typography components to core-elements
