+++ packages/author/package.json
@@ -16,7 +16,6 @@
   },
   "dependencies": {
     "@titicaca/color-palette": "^2.0.0-alpha.1",
-    "@titicaca/core-elements": "^2.0.0-alpha.1",
-    "@titicaca/triple-document": "^2.0.0-alpha.1"
+    "@titicaca/core-elements": "^2.0.0-alpha.1"
   }
 }
