+++ packages/core-elements/src/index.ts
@@ -11,7 +11,8 @@ export {
 export { default as Radio } from './elements/radio'
 export { default as GenderSelector } from './elements/gender-selector'
 export { default as Section } from './elements/section'
-export { default as Text, TextProps, TextBaseProps } from './elements/text'
+export { default as Text, TextProps } from './elements/text'
+export * from './elements/typography'
 export { default as Tag, TagColors } from './elements/tag'
 export { default as Label, LabelColor } from './elements/label'
 export { default as Icon } from './elements/icon'
