+++ packages/nearby-pois/package.json
@@ -20,7 +20,6 @@
     "@titicaca/intersection-observer": "^2.0.0-alpha.1",
     "@titicaca/poi-list-elements": "^2.0.0-alpha.1",
     "@titicaca/react-contexts": "^2.0.0-alpha.1",
-    "@titicaca/triple-document": "^2.0.0-alpha.1",
     "@titicaca/type-definitions": "^2.0.0-alpha.1",
     "@titicaca/view-utilities": "^2.0.0-alpha.1",
     "isomorphic-fetch": "^2.2.1"
