+++ packages/poi-detail/package.json
@@ -22,7 +22,6 @@
     "@titicaca/intersection-observer": "^2.0.0-alpha.1",
     "@titicaca/modals": "^2.0.0-alpha.1",
     "@titicaca/react-contexts": "^2.0.0-alpha.1",
-    "@titicaca/triple-document": "^2.0.0-alpha.1",
     "@titicaca/type-definitions": "^2.0.0-alpha.1",
     "@titicaca/view-utilities": "^2.0.0-alpha.1",
     "isomorphic-fetch": "^2.2.1"
