+++ packages/triple-document/package.json
@@ -21,6 +21,7 @@
     "@titicaca/modals": "^2.0.0-alpha.1",
     "@titicaca/poi-list-elements": "^2.0.0-alpha.1",
     "@titicaca/react-contexts": "^2.0.0-alpha.1",
+    "@titicaca/standard-action-handler": "^2.0.0-alpha.1",
     "@titicaca/triple-media": "^2.0.0-alpha.1",
     "@titicaca/type-definitions": "^2.0.0-alpha.1",
     "@titicaca/view-utilities": "^2.0.0-alpha.1",
