+++ packages/triple-document/src/elements/text/headings.tsx
@@ -1,86 +1,31 @@
 import * as React from 'react'
-import {
-  Container,
-  Text,
-  TextProps,
-  MarginPadding,
-} from '@titicaca/core-elements'
-import * as CSS from 'csstype'
+import { H1, H2, H3, H4 } from '@titicaca/core-elements'
 
-export interface H1Props extends TextProps {
+interface HeadingProps {
   href?: string
-  headline?: string
   emphasize?: boolean
-  margin?: MarginPadding
-  textAlign?: CSS.Property.TextAlign
-  children?: string
+  headline?: string
+  children: string
 }
 
-export type H2Props = TextProps
-export type H3Props = TextProps
-export type H4Props = TextProps
-
-export function H1({
-  href,
-  headline,
-  emphasize,
-  margin,
-  textAlign,
-  children,
-  ...props
-}: H1Props) {
-  return (
-    <Container id={href} margin={margin} textAlign={textAlign}>
-      {headline && (
-        <Text bold size="tiny" color="blue" margin={{ bottom: 3 }}>
-          {headline}
-        </Text>
-      )}
-      <Text bold size="huge" color={emphasize ? 'blue' : 'gray'} {...props}>
-        {children}
-      </Text>
-    </Container>
-  )
-}
-
-export function H2({ children, ...props }: H2Props) {
-  return (
-    <Text size="big" color="gray" {...props}>
+export const MH1 = tripleDocumentHeading(
+  ({ children, ...props }: HeadingProps) => (
+    <H1 margin={{ top: 25, bottom: 20, left: 30, right: 30 }} {...props}>
       {children}
-    </Text>
-  )
-}
+    </H1>
+  ),
+)
 
-export function H3({ children, ...props }: H3Props) {
-  return (
-    <Text bold size="large" color="gray" {...props}>
+export const MH2 = tripleDocumentHeading(
+  ({ children, ...props }: HeadingProps) => (
+    <H2 margin={{ top: 20, bottom: 20, left: 30, right: 30 }} {...props}>
       {children}
-    </Text>
-  )
-}
-
-export function H4({ children, ...props }: H4Props) {
-  return (
-    <Text bold size="large" color="blue" {...props}>
-      {children}
-    </Text>
-  )
-}
-
-export const MH1 = tripleDocumentHeading(({ children, ...props }: H1Props) => (
-  <H1 margin={{ top: 25, bottom: 20, left: 30, right: 30 }} {...props}>
-    {children}
-  </H1>
-))
-
-export const MH2 = tripleDocumentHeading(({ children, ...props }: H2Props) => (
-  <H2 margin={{ top: 20, bottom: 20, left: 30, right: 30 }} {...props}>
-    {children}
-  </H2>
-))
+    </H2>
+  ),
+)
 
 export const MH3 = tripleDocumentHeading(
-  ({ compact, children, ...props }: H3Props & { compact: boolean }) => (
+  ({ compact, children, ...props }: HeadingProps & { compact: boolean }) => (
     <H3
       margin={compact ? { top: 13 } : { top: 20, left: 30, right: 30 }}
       {...props}
@@ -90,20 +35,16 @@ export const MH3 = tripleDocumentHeading(
   ),
 )
 
-export const MH4 = tripleDocumentHeading(({ children, ...props }: H4Props) => (
-  <H4 margin={{ top: 20, left: 30, right: 30 }} {...props}>
-    {children}
-  </H4>
-))
+export const MH4 = tripleDocumentHeading(
+  ({ children, ...props }: HeadingProps) => (
+    <H4 margin={{ top: 20, left: 30, right: 30 }} {...props}>
+      {children}
+    </H4>
+  ),
+)
 
 function tripleDocumentHeading<P extends object>(
-  Component: React.ComponentType<
-    P & {
-      href: string
-      emphasize: boolean
-      headline: string
-    }
-  >,
+  Component: React.ComponentType<P & HeadingProps>,
 ) {
   return function WrappedHeading({
     value: { text, href, emphasize, headline },
