+++ packages/triple-document/src/elements/text/index.ts
@@ -1,3 +1,2 @@
 export * from './headings'
 export { default as Text } from './plain'
-export { default as Paragraph } from './paragraph'
